+++ code/src/sixsq/nuvla/server/resources/nuvlabox/status_utils.clj
@@ -56,9 +56,7 @@
     response))
 
 (defn detect-swarm
-  [{{:keys [parent orchestrator node-id cluster-node-role cluster-managers swarm-node-id]} :body :as _response}
-   _request]
-
+  [{{:keys [parent orchestrator node-id cluster-node-role cluster-managers swarm-node-id] :as _nuvlabox-status} :body :as _response}]
   (let [{:keys [infrastructure-service-group] :as _nuvlabox} (db/retrieve parent)
         swarm-node-id-set? (not (str/blank? swarm-node-id))
         attributes         {:swarm-enabled (or (= "swarm" orchestrator)
@@ -70,9 +68,8 @@
     (log/debugf "detect-swarm - parent: %s - isg: %s - attrs: %s - swarm-node-id: %s - orchestrator: %s - node-id: %s - cluster-node-role: %s - cluster-managers: %s"
                 parent infrastructure-service-group attributes swarm-node-id orchestrator node-id cluster-node-role cluster-managers)
     (when-let [resource-id (nb-utils/get-service "swarm" infrastructure-service-group)]
-      (log/debugf "detect-swarm - parent: %s - resource-id: %s - scripted-edit: %s"
-                  parent resource-id (db/scripted-edit resource-id {:refresh false
-                                                                    :body    {:doc attributes}})))))
+      (let [response (db/scripted-edit resource-id {:refresh false :body {:doc attributes}})]
+        (log/debugf "detect-swarm - parent: %s - resource-id: %s - scripted-edit: %s" parent resource-id response)))))
 
 (defmulti nuvlabox-status->metric-data (fn [_ metric] metric))