+++ code/src/sixsq/nuvla/server/util/general.clj
@@ -1,4 +1,8 @@
-(ns sixsq.nuvla.server.util.general)
+(ns sixsq.nuvla.server.util.general
+  (:require [clojure.string :as str])
+  (:import
+    (java.net URLEncoder, URLDecoder)
+    (java.nio.charset StandardCharsets)))
 
 (defn filter-map-nil-value
   [m]
@@ -10,3 +14,29 @@
 
 (defn index-by [coll key-fn]
   (into {} (map (juxt key-fn identity) coll)))
+
+(defn replace-in-str
+  [s r-map]
+  (reduce-kv (fn [in re v] (str/replace in re v)) s r-map))
+
+(defn encode-uri-component
+  [^String s]
+  (-> s
+      (URLEncoder/encode StandardCharsets/UTF_8)
+      (replace-in-str {#"\+"  "%20"
+                       #"%21" "!"
+                       #"%27" "'"
+                       #"%28" "("
+                       #"%29" ")"
+                       #"%7E" "~"})))
+
+(defn decode-uri-component
+  [^String s]
+  (-> s
+      ^String (replace-in-str {#"%20" "+"
+                               #"\!"  "%21"
+                               #"\'"  "%27"
+                               #"\("  "%28"
+                               #"\)"  "%29"
+                               #"\~"  "%7E"})
+      (URLDecoder/decode StandardCharsets/UTF_8)))