+++ code/test/sixsq/nuvla/server/resources/event/test_utils.clj
@@ -3,9 +3,9 @@
     [clojure.data.json :as json]
     [clojure.string :as str]
     [peridot.core :refer [content-type header request session]]
-    [ring.util.codec :as rc]
     [sixsq.nuvla.server.middleware.authn-info :refer [authn-info-header]]
     [sixsq.nuvla.server.resources.lifecycle-test-utils :as ltu]
+    [sixsq.nuvla.server.util.general :as gen-util]
     [sixsq.nuvla.server.util.time :as time]))
 
 
@@ -14,7 +14,7 @@
   (->> (re-seq #"([^=]*)=(.*)" p)
        first
        next
-       (map rc/url-encode)
+       (map gen-util/encode-uri-component)
        (str/join "=")))