+++ code/test/sixsq/nuvla/server/resources/notification_method_test.clj
@@ -13,7 +13,9 @@
 
 (def base-uri (str p/service-context t/resource-type))
 
-(def valid-notif-conf {:method      "email"
+(def valid-notif-conf {:name "Jane's email"
+                       :description "Jane's email"
+                       :method      "email"
                        :destination "jane@example.com"
                        :acl         {:owners ["user/jane"]}})
 
@@ -92,3 +94,59 @@
           (request user-abs-uri :request-method :delete)
           (ltu/body->edn)
           (ltu/is-status 200)))))
+
+(deftest action-test
+  (let [session-anon (-> (ltu/ring-app)
+                         session
+                         (content-type "application/json"))
+        session-user (header session-anon authn-info-header "user/jane user/jane group/nuvla-user group/nuvla-anon")]
+
+    (let [user-uri (-> session-user
+                       (request base-uri
+                                :request-method :post
+                                :body (json/write-str valid-notif-conf))
+                       (ltu/body->edn)
+                       (ltu/is-status 201)
+                       (ltu/location))
+
+          user-abs-uri (str p/service-context user-uri)]
+
+      (-> session-user
+          (request base-uri)
+          (ltu/body->edn)
+          (ltu/is-status 200)
+          (ltu/is-resource-uri t/collection-type)
+          (ltu/is-count 1))
+
+      (let [test-op (-> session-user
+                        (request user-abs-uri)
+                        (ltu/body->edn)
+                        (ltu/is-status 200)
+                        (ltu/is-operation-present :test)
+                        (ltu/get-op-url :test))]
+        (-> session-user
+            (request test-op
+                     :request-method :post
+                     :body (json/write-str {}))
+            (ltu/body->edn)
+            (ltu/is-status 201)
+            (ltu/message-matches t/test-response-message)))
+
+      (ltu/refresh-es-indices)
+
+      (let [event (-> session-user
+                      (request "/api/event")
+                      (ltu/body->edn)
+                      (ltu/is-count 1)
+                      (ltu/entries)
+                      first)
+            resource-content (-> event :content :resource :content)]
+        (is (= "test.notification" (:name event)))
+        (is (= (:method valid-notif-conf) (:method resource-content)))
+        (is (= (:destination valid-notif-conf) (:destination resource-content)))
+
+      ;; user can delete
+      (-> session-user
+          (request user-abs-uri :request-method :delete)
+          (ltu/body->edn)
+          (ltu/is-status 200))))))